+++ simplesettings/src/main/java/com/chillibits/simplesettings/core/elements/PreferenceHeader.kt
@@ -1,4 +1,4 @@
-package com.chillibits.simplesettings.core
+package com.chillibits.simplesettings.core.elements
 
 import android.content.Context
 import androidx.annotation.LayoutRes
