+++ simplesettings/src/main/java/com/chillibits/simplesettings/core/elements/PreferencePage.kt
@@ -2,9 +2,10 @@
  * Copyright © Marc Auberer 2020-2021. All rights reserved
  */
 
-package com.chillibits.simplesettings.core
+package com.chillibits.simplesettings.core.elements
 
 import android.content.Context
+import com.chillibits.simplesettings.core.SimpleSettings
 import com.chillibits.simplesettings.item.SimplePreference
 
 /**
