+++ simplesettings/src/main/java/com/chillibits/simplesettings/ui/SimpleSettingsActivity.kt
@@ -13,9 +13,9 @@ import android.view.MenuItem
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import com.chillibits.simplesettings.R
-import com.chillibits.simplesettings.core.PreferenceHeader
-import com.chillibits.simplesettings.core.PreferenceSection
 import com.chillibits.simplesettings.core.SimpleSettings
+import com.chillibits.simplesettings.core.elements.PreferenceHeader
+import com.chillibits.simplesettings.core.elements.PreferenceSection
 import com.chillibits.simplesettings.exception.SettingsResetException
 import com.chillibits.simplesettings.tool.Constants
 import com.chillibits.simplesettings.tool.getPrefs
